python-backend/requirements.txt
fastapi 
uvicorn 
pinecone
sentence-transformers
httpx
python-dotenv
numpy==1.26.4